refactor: create GetUserProfileGateway and GetUserProfileUseCase
