+++ src/@core/frontend/infra/gateways/user/httpUserGateway.ts
@@ -3,12 +3,8 @@ import {
     CreateUserGateway,
     SignupData
 } from "@/@core/frontend/domain/gateways/user/createUserGateway"
-import {
-    GetUserGateway,
-    UserOrGuestToken
-} from "@/@core/frontend/domain/gateways/user/getUserGateway"
 
-export class HttpUserGateway implements CreateUserGateway, GetUserGateway {
+export class HttpUserGateway implements CreateUserGateway {
     public async create(userInfo: SignupData): Promise<boolean> {
         const headers: HeadersInit = { "Content-type": "application/json" }
         const response = await fetch("/api/signup", {
@@ -29,24 +25,5 @@ export class HttpUserGateway implements CreateUserGateway, GetUserGateway {
         }
 
         return true
-    }
-
-    public async getUser(): Promise<UserOrGuestToken> {
-        const accessToken = localStorage.getItem("accessToken") as string
-
-        const response = await fetch("/api/auth/user", {
-            headers: { Authorization: `Bearer ${accessToken}` }
-        })
-
-        const responseData = await response.json()
-
-        if (!response.ok && responseData.errors) {
-            const { errors } = responseData as HttpGatewayResponse<"failed">
-            throw new Error(errors.join(","))
-        }
-
-        const { data } = responseData as HttpGatewayResponse<"success", UserOrGuestToken>
-
-        return data
     }
 }
