+++ src/@core/frontend/main/gateways/userProfileGatewayFactory.ts
@@ -1,0 +1,7 @@
+import { HttpUserProfileGateway } from "../../infra/gateways/user/httpUserProfileGateway"
+
+function createHttpUserProfileGateway(): HttpUserProfileGateway {
+    return new HttpUserProfileGateway()
+}
+
+export const httpUserProfileGateway = createHttpUserProfileGateway()
