--- src/@core/frontend/main/usecases/user/getUserFactory.ts
@@ -1,11 +1,0 @@
-import { HttpUserGateway } from "@/@core/frontend/infra/gateways/user/httpUserGateway"
-import { GetUserUseCase } from "@/@core/frontend/usecases/user/getUserUseCase"
-
-function createGetUserUseCase(): GetUserUseCase {
-    const httpUserGateway = new HttpUserGateway()
-    const getUserUseCase = new GetUserUseCase(httpUserGateway)
-
-    return getUserUseCase
-}
-
-export const getUserUseCase = createGetUserUseCase()
