+++ src/@core/frontend/main/usecases/user/getUserProfileFactory.ts
@@ -1,0 +1,8 @@
+import { httpUserProfileGateway } from "../../gateways/userProfileGatewayFactory"
+import { GetUserUseProfileUseCase } from "@/@core/frontend/usecases/user/profile/getUserProfileUseCase"
+
+function createGetUserProfileUseCase(): GetUserUseProfileUseCase {
+    return new GetUserUseProfileUseCase(httpUserProfileGateway)
+}
+
+export const getUserProfileUseCase = createGetUserProfileUseCase()
