+++ src/@core/frontend/usecases/user/profile/getUserProfileUseCase.ts
@@ -1,0 +1,17 @@
+import {
+    DefaultUser,
+    GetUserProfileGateway,
+    GuestUser
+} from "../../../domain/gateways/user/profile/getUserProfileGateway"
+
+type GetUserOutputProfileDTO = DefaultUser | GuestUser | string
+
+export class GetUserUseProfileUseCase {
+    constructor(private readonly getUserProfileGateway: GetUserProfileGateway) {}
+
+    public async execute(): Promise<GetUserOutputProfileDTO> {
+        const foundUser = await this.getUserProfileGateway.getProfile()
+
+        return foundUser
+    }
+}
